rename: 컨트롤 관련 컴퍼넌트 Control 폴더에 통합 #50
* /components/ -> /components/Control/
